refactor: Move utility defensive checkers to dedicated module
